Factor routes and their integration tests
