+++ frontend/src/components/JoinGroupForm.jsx
@@ -22,7 +22,7 @@ const JoinGroupForm = () => {
             setIsValid(false)
             return
         }
-        fetch(`/api/group/${groupToken}`, {
+        fetch(`/api/groups/${groupToken}`, {
             method: 'GET',
             headers: { 'Content-Type': 'application/json' },
         })
